Move AttentionMetadataBuilder and AttentionCGSupport from utils.py to backend.py
